Fixed longFormatter in BasicFormatters (incorrect byteValue call was used to parse value from a string).

+++ src/main/java/net/formio/format/BasicFormatters.java
@@ -196,7 +196,7 @@ public class BasicFormatters implements Formatters {
 					try {
 						String amendedStr = removeDecimalPart(str, loc.getLocale());
 						return Long.valueOf(FormatsCache.getOrCreateDecimalFormat(formatPattern, loc)
-							.parse(amendedStr).byteValue());
+							.parse(amendedStr).longValue());
 					} catch (Exception ex) {
 						throw new StringParseException(Long.class, str, ex);
 					}
